+++ src/pages/TradeReport.jsx
@@ -1,92 +1,589 @@
-import React, { Component } from 'react';
 
+import React, { useState, useEffect } from "react";
+import { createPortal } from "react-dom";
 
-import ComboBox from '../componentes/ComboBox'
-import ComboBox2 from '../componentes/ComboBox2'
-import ComboBox3 from '../componentes/ComboBox3'
-
-import Dashboard from '../componentes/Dashboard'
-import TabSelector from '../componentes/TabSelector'
-
+import BarChart from "../componentes/BarChart";
+import ComboBoxTradeReportersImporters from "../componentes/ComboBoxTradeReporters";
+import CountryCharacteristics from '../data/CountryCharacteristics.json';
 import DashboardTradeReport from '../componentes/DashboardTradeReport'
-//dashboards
-import SustainableImporters from './SustainableImporters'
-import SustainableExporters from './SustainableExporters'
 
 
+const SustainableImporters =()=>
+ {
+  
+  const [state,setState]=useState({select: {
+  Product: 'abaca',
+ iteration: "4",
+ scenathon_id :"6",
+ column:"Export_quantity"
+ }});
 
+  //const [state,setState]=useState([]);
+  const [json,setJson]=useState([]);
+  var dataAux = null;
 
+ const handleChange = e => {
+var scenathon_id=state.select.scenathon_id;
+var columnAux=state.select.column;
+var Product =state.select.Product;
+var iteration=state.select.iteration;
 
-class TradeReport extends Component {
+ 
+  if(e.target.name=="dashboard-tradeReport"){
+  switch (e.target.value) {
+    case 'Sustainable_next_exporters':
+       // this.dash = <SustainableImporters scenathon_id={"6"} column={"Export_quantity"} iteration={"4"}/>;
+       scenathon_id="6";
+       columnAux="Export_quantity";
+      iteration=state.select.iteration=="1"? "3":"4";
+       
+        break;
+    case 'Sustainable_next_importers':
+     //   this.dash = <SustainableImporters scenathon_id={"5"} column={"Import_quantity"} iteration={"4"}/>;
+     scenathon_id="6";
+     columnAux="Import_quantity";
+     iteration=state.select.iteration=="1"? "3":"4";
+        break;
+    case 'Current_trend_next_exporters':
+      //  this.dash = <h1>Current_trend_next_exporters</h1>
+      scenathon_id="5";
+      columnAux="Export_quantity";
+     iteration=state.select.iteration=="3"? "1":"2";
+//iteration="2";
+        break;
+    case 'Current_trend_next_importers':
+      scenathon_id="5";
+      columnAux="Import_quantity";
+      iteration=state.select.iteration=="3"? "1":"2";
+      //  this.dash = <h1>Current_trend_next_importers</h1>
+        break;
 
-
-    state = {
-        select: {
-            dashboard: "Sustainable_next_exporters"
-        }
-
-    }
-
-
-    handleChange = e => {
-
-        this.setState({
-            select: {
-                //el next code evitara que se sobrescriba cuando reciba un valor new
-                ...this.state.select.dashboard,
-                ["dashboard"]: e.target.value
-
-            }
-
-        })
-
-
-
-    }
-
-    selectDashboard(state) {
-
-        console.log("entre switch")
-        switch (state) {
-            case 'Sustainable_next_exporters':
-                this.dash = <SustainableImporters scenathon_id={"6"} column={"Export_quantity"} iteration={"4"}/>;
-                console.log("entre primero")
-                break;
-            case 'Sustainable_next_importers':
-                this.dash = <SustainableImporters scenathon_id={"5"} column={"Import_quantity"} iteration={"4"}/>;
-                console.log("entre 2do")
-                break;
-            case 'Current_trend_next_exporters':
-                this.dash = <h1>Current_trend_next_exporters</h1>
-                break;
-            case 'Current_trend_next_importers':
-                this.dash = <h1>Current_trend_next_importers</h1>
-                break;
-
-                
-        }
-
-    }
-
-    render() {
-        return (
-            <div className="container">
-                <div>
-                    {/* mostrar las opciones de los diferentes tipos de dashboards en trade report*/}
-                    <div className="dashboards-tradeReport">
-                        <DashboardTradeReport metodo={this.handleChange} />
-                    </div>
-
-                    {/* metodo para saber que dashboard se elijio*/}
-                    {this.selectDashboard(this.state.select.dashboard)}
-
-                    {/*mostrar el dashboard elegido */}
-                    {this.dash}
-
-                </div>
-            </div>
-        )
-    }
+        
 }
-export default TradeReport;
+}else{
+  
+  Product= e.target.name=="Product"? e.target.value: state.select.Product;
+  iteration=e.target.name=="iteration"?scenathon_id==="6" ? e.target.value==="after"? "4":"3" : e.target.value==="after"? "2":"1":state.select.iteration;
+}
 
+setState({
+      select: {
+        Product: Product,
+        iteration: iteration,
+        scenathon_id :scenathon_id,
+        column:columnAux
+        
+      }
+
+});
+
+
+
+  }
+
+  
+
+  useEffect(() => 
+  {
+    
+  getNettrade();
+  }, [state]);
+
+ 
+
+  const getNettrade = async() => {
+    try {
+   
+      const body =state;
+      
+   
+      const response = await fetch("http://localhost:5000/net/"+JSON.stringify(body));
+     const  jsonAux =  await response.json();
+    
+    setJson(jsonAux);
+
+    } catch (error) {
+      console.error(error)
+    }
+
+
+
+  }
+
+  function Pais(CountryCharacteristics,data) {
+    this.data=data;
+    this.type=CountryCharacteristics[0]["type"];
+    this.label=CountryCharacteristics[0]["label"];
+    this.borderColor=CountryCharacteristics[0]["borderColor"];
+    this.backgroundColor=CountryCharacteristics[0]["backgroundColor"];
+    
+  }
+
+const getImportQuantity=()=>
+{
+  var contador=0;
+var data=[];
+var paises=[];
+var labels=[];
+
+  if (json != null) {
+    json.map((item) => {
+      if (!labels.includes(item.Year)) 
+      {
+        labels.push(item.Year);
+      }
+      data.push(item.Import_quantity);
+      contador=contador+1;
+      if (contador==11) {
+        var pais = new Pais(CountryCharacteristics[item.name], data);
+          paises.push(pais);
+          contador=0;
+        data=[];
+      }
+    });
+
+
+  }
+  data = {
+    labels:labels,
+    datasets:paises
+};
+ dataAux=data;
+}
+
+const getExportQuantity=()=>
+{
+  var contador=0;
+var data=[];
+var paises=[];
+var labels=[];
+
+  if (json != null) {
+    json.map((item) => {
+      if (!labels.includes(item.Year)) 
+      {
+        labels.push(item.Year);
+      }
+      data.push(item.Export_quantity);
+      contador=contador+1;
+      if (contador==11) {
+        var pais = new Pais(CountryCharacteristics[item.name], data);
+          paises.push(pais);
+          contador=0;
+        data=[];
+      }
+    });
+
+
+  }
+  data = {
+    labels:labels,
+    datasets:paises
+};
+ dataAux=data;
+}
+
+
+
+const converter=()=>{
+
+
+if(state.select.column=="Export_quantity")
+{
+  getExportQuantity();
+}else{
+ 
+  getImportQuantity();
+}
+
+
+ 
+
+
+ 
+
+  }   
+
+ 
+    return (
+
+      <div>
+        {converter()}
+       
+        <div>
+        <DashboardTradeReport metodo={handleChange} />
+           <ComboBoxTradeReportersImporters metodo={handleChange} />
+        
+        </div>
+
+        {/*this.selectDashboard()*/}
+       
+
+        <div>
+          <BarChart data={dataAux} title="Sustainable - net exporters" />
+        </div>
+
+
+      </div>
+
+
+    )
+  
+}
+
+
+const convertir = (props) => {
+  var CHN = [];
+  var R_OEU = [];
+  var R_CSA = [];
+  var R_NEU = [];
+  var ARG = [];
+  var ZAF = [];
+  var BRA = [];
+  var IND = [];
+  var CAN = [];
+  var USA = []
+  var AUS = [];
+  var R_SSA = [];
+  var DEU = [];
+  var NOR = [];
+  var MEX = [];
+  var COL = [];
+  var FIN = [];
+  var MYS = [];
+  var R_NMC = [];
+  var R_ASP = [];
+  var GBR = [];
+
+  //se guarda el anio
+  var labels = [];
+
+
+  if (props != undefined) {
+    props.map((item) => {
+      //export quantity (unit 1000 tons) was replace to export_quantity 
+
+      if (item.alpha2 === "CHN") {
+        CHN.push(item.export_quantity);
+      } else if (item.alpha2 === "R_OEU") {
+        R_OEU.push(item.export_quantity);
+      } else if (item.alpha2 === "R_CSA") {
+        R_CSA.push(item.export_quantity);
+      } else if (item.alpha2 === "R_NEU") {
+        R_NEU.push(item.export_quantity);
+      }
+      else if (item.alpha2 === "ARG") {
+        ARG.push(item.export_quantity);
+      } else if (item.alpha2 === "ZAF") {
+        ZAF.push(item.export_quantity);
+      } else if (item.alpha2 === "BRA") {
+        BRA.push(item.export_quantity);
+      } else if (item.alpha2 === "IND") {
+        IND.push(item.export_quantity);
+      } else if (item.alpha2 === "CAN") {
+        CAN.push(item.export_quantity);
+      } else if (item.alpha2 === "USA") {
+        USA.push(item.export_quantity);
+      } else if (item.alpha2 === "AUS") {
+        AUS.push(item.export_quantity);
+      } else if (item.alpha2 === "R_SSA") {
+        R_SSA.push(item.export_quantity);
+      } else if (item.alpha2 === "DEU") {
+        DEU.push(item.export_quantity);
+      } else if (item.alpha2 === "NOR") {
+        NOR.push(item.export_quantity);
+      } else if (item.alpha2 === "MEX") {
+        MEX.push(item.export_quantity);
+      } else if (item.alpha2 === "COL") {
+        COL.push(item.export_quantity);
+      } else if (item.alpha2 === "FIN") {
+        FIN.push(item.export_quantity);
+      } else if (item.alpha2 === "MYS") {
+        MYS.push(item.export_quantity);
+      } else if (item.alpha2 === "R_NMC") {
+        R_NMC.push(item.export_quantity);
+      } else if (item.alpha2 === "R_ASP") {
+        R_ASP.push(item.export_quantity);
+      } else if (item.alpha2 === "GBR") {
+        GBR.push(item.export_quantity);
+      }
+
+      if (!labels.includes(item.year)) {
+        labels.push(item.year);
+      }
+    }
+
+
+    );
+  }
+
+  const data = {
+    labels: labels,
+    datasets: [
+      {
+        //Aforestation ejemplo
+        type: 'bar',
+        label: 'CHN',
+        data: CHN,
+        fill: false,
+        backgroundColor: '#faa75a',
+        borderColor: '#71B37C',
+        hoverBackgroundColor: '#71B37C',
+        hoverBorderColor: '#71B37C',
+        yAxisID: 'y-axis-1'
+      }, {
+        //Aforestation ejemplo
+        type: 'bar',
+        label: 'GBR',
+        data: GBR,
+        fill: false,
+        backgroundColor: '#faa75a',
+        borderColor: '#71B37C',
+        hoverBackgroundColor: '#71B37C',
+        hoverBorderColor: '#71B37C',
+        yAxisID: 'y-axis-1'
+      }, {
+        //Aforestation ejemplo
+        type: 'bar',
+        label: 'R_ASP',
+        data: R_ASP,
+        fill: false,
+        backgroundColor: '#faa75a',
+        borderColor: '#71B37C',
+        hoverBackgroundColor: '#71B37C',
+        hoverBorderColor: '#71B37C',
+        yAxisID: 'y-axis-1'
+      }, {
+        type: 'bar',
+        label: 'R_OEU',
+        data: R_OEU,
+        fill: false,
+        borderColor: '#EC932F',
+        backgroundColor: '#7cb342',
+        pointBorderColor: '#EC932F',
+        pointBackgroundColor: '#EC932F',
+        pointHoverBackgroundColor: '#EC932F',
+        pointHoverBorderColor: '#EC932F',
+        yAxisID: 'y-axis-1'
+      }, {
+        type: 'bar',
+        label: 'R_CSA',
+        data: R_CSA,
+        fill: false,
+        borderColor: '#EC932F',
+        backgroundColor: '#5a9bd4',
+        pointBorderColor: '#EC932F',
+        pointBackgroundColor: '#EC932F',
+        pointHoverBackgroundColor: '#EC932F',
+        pointHoverBorderColor: '#EC932F',
+        yAxisID: 'y-axis-1'
+      },
+      {
+        type: 'bar',
+        label: 'R_NEU',
+        data: R_NEU,
+        fill: false,
+        borderColor: '#EC932F',
+        backgroundColor: '#5e35b1',
+        pointBorderColor: '#EC932F',
+        pointBackgroundColor: '#EC932F',
+        pointHoverBackgroundColor: '#EC932F',
+        pointHoverBorderColor: '#EC932F',
+        yAxisID: 'y-axis-1'
+      },
+      {
+        type: 'bar',
+        label: 'ARG',
+        data: ARG,
+        fill: false,
+        borderColor: '#EC932F',
+        backgroundColor: '#00b6cb',
+        pointBorderColor: '#EC932F',
+        pointBackgroundColor: '#EC932F',
+        pointHoverBackgroundColor: '#EC932F',
+        pointHoverBorderColor: '#EC932F',
+        yAxisID: 'y-axis-1'
+      }
+      ,
+      {
+        type: 'bar',
+        label: 'ZAF',
+        data: ZAF,
+        fill: false,
+        borderColor: '#EC932F',
+        backgroundColor: '#5e35b1',
+        pointBorderColor: '#EC932F',
+        pointBackgroundColor: '#EC932F',
+        pointHoverBackgroundColor: '#EC932F',
+        pointHoverBorderColor: '#EC932F',
+        yAxisID: 'y-axis-1'
+      },
+      {
+        type: 'bar',
+        label: 'BRA',
+        data: BRA,
+        fill: false,
+        borderColor: '#EC932F',
+        backgroundColor: '#00b6cb',
+        pointBorderColor: '#EC932F',
+        pointBackgroundColor: '#EC932F',
+        pointHoverBackgroundColor: '#EC932F',
+        pointHoverBorderColor: '#EC932F',
+        yAxisID: 'y-axis-1'
+      },
+      {
+        type: 'bar',
+        label: 'IND',
+        data: IND,
+        fill: false,
+        borderColor: '#EC932F',
+        backgroundColor: '#0072f0',
+        pointBorderColor: '#EC932F',
+        pointBackgroundColor: '#EC932F',
+        pointHoverBackgroundColor: '#EC932F',
+        pointHoverBorderColor: '#EC932F',
+        yAxisID: 'y-axis-1'
+      }, {
+        type: 'bar',
+        label: 'CAN',
+        data: CAN,
+        fill: false,
+        borderColor: '#EC932F',
+        backgroundColor: '#9e67ab',
+        pointBorderColor: '#EC932F',
+        pointBackgroundColor: '#EC932F',
+        pointHoverBackgroundColor: '#EC932F',
+        pointHoverBorderColor: '#EC932F',
+        yAxisID: 'y-axis-1'
+      }, {
+        type: 'bar',
+        label: 'USA',
+        data: USA,
+        fill: false,
+        borderColor: '#EC932F',
+        backgroundColor: '#0072f0',
+        pointBorderColor: '#EC932F',
+        pointBackgroundColor: '#EC932F',
+        pointHoverBackgroundColor: '#EC932F',
+        pointHoverBorderColor: '#EC932F',
+        yAxisID: 'y-axis-1'
+      },
+      {
+        type: 'bar',
+        label: 'AUS',
+        data: AUS,
+        fill: false,
+        borderColor: '#EC932F',
+        backgroundColor: '##ce7058',
+        pointBorderColor: '#EC932F',
+        pointBackgroundColor: '#EC932F',
+        pointHoverBackgroundColor: '#EC932F',
+        pointHoverBorderColor: '#EC932F',
+        yAxisID: 'y-axis-1'
+      },
+      {
+        type: 'bar',
+        label: 'R_SSA',
+        data: R_SSA,
+        fill: false,
+        borderColor: '#EC932F',
+        backgroundColor: '#ffa800',
+        pointBorderColor: '#EC932F',
+        pointBackgroundColor: '#EC932F',
+        pointHoverBackgroundColor: '#EC932F',
+        pointHoverBorderColor: '#EC932F',
+        yAxisID: 'y-axis-1'
+      }, {
+        type: 'bar',
+        label: 'DEU',
+        data: DEU,
+        fill: false,
+        borderColor: '#EC932F',
+        backgroundColor: '#ec407a',
+        pointBorderColor: '#EC932F',
+        pointBackgroundColor: '#EC932F',
+        pointHoverBackgroundColor: '#EC932F',
+        pointHoverBorderColor: '#EC932F',
+        yAxisID: 'y-axis-1'
+      }, {
+        type: 'bar',
+        label: 'NOR',
+        data: NOR,
+        fill: false,
+        borderColor: '#EC932F',
+        backgroundColor: '#7ac36a',
+        pointBorderColor: '#EC932F',
+        pointBackgroundColor: '#EC932F',
+        pointHoverBackgroundColor: '#EC932F',
+        pointHoverBorderColor: '#EC932F',
+        yAxisID: 'y-axis-1'
+      }, {
+        type: 'bar',
+        label: 'MEX',
+        data: MEX,
+        fill: false,
+        borderColor: '#EC932F',
+        backgroundColor: '#f10096',
+        pointBorderColor: '#EC932F',
+        pointBackgroundColor: '#EC932F',
+        pointHoverBackgroundColor: '#EC932F',
+        pointHoverBorderColor: '#EC932F',
+        yAxisID: 'y-axis-1'
+      },
+      {
+        type: 'bar',
+        label: 'COL',
+        data: COL,
+        fill: false,
+        borderColor: '#EC932F',
+        backgroundColor: '#ffa800',
+        pointBorderColor: '#EC932F',
+        pointBackgroundColor: '#EC932F',
+        pointHoverBackgroundColor: '#EC932F',
+        pointHoverBorderColor: '#EC932F',
+        yAxisID: 'y-axis-1'
+      },
+      {
+        type: 'bar',
+        label: 'FIN',
+        data: FIN,
+        fill: false,
+        borderColor: '#EC932F',
+        backgroundColor: '#5e35b1',
+        pointBorderColor: '#EC932F',
+        pointBackgroundColor: '#EC932F',
+        pointHoverBackgroundColor: '#EC932F',
+        pointHoverBorderColor: '#EC932F',
+        yAxisID: 'y-axis-1'
+      }, {
+        type: 'bar',
+        label: 'MYS',
+        data: MYS,
+        fill: false,
+        borderColor: '#EC932F',
+        backgroundColor: '#0390d0',
+        pointBorderColor: '#EC932F',
+        pointBackgroundColor: '#EC932F',
+        pointHoverBackgroundColor: '#EC932F',
+        pointHoverBorderColor: '#EC932F',
+        yAxisID: 'y-axis-1'
+      }, {
+        type: 'bar',
+        label: 'R_NMC',
+        data: R_NMC,
+        fill: false,
+        borderColor: '#EC932F',
+        backgroundColor: '#f10096',
+        pointBorderColor: '#EC932F',
+        pointBackgroundColor: '#EC932F',
+        pointHoverBackgroundColor: '#EC932F',
+        pointHoverBorderColor: '#EC932F',
+        yAxisID: 'y-axis-1'
+      }
+    ]
+  };
+
+  return data
+}
+
+export default SustainableImporters;
